Crate air_log_targets

Constants

DATA_CACHE

Print out data cache at the beginning of each instruction execution_step.

EXECUTED_STATE_CHANGING

Print out state of data cache at the beginning of each instruction execution_step.

EXECUTED_TRACE_MERGE

Print out logs at the executed states merging stage.

INSTRUCTION

Print out each instruction name at the beginning of its execution_step.

JOIN_BEHAVIOUR

Print log if call is postponed due the join behaviour.

NEW_EXECUTED_TRACE

Print out state of data cache at the beginning of each instruction execution_step.

NEXT_PEER_PKS

Print out next_peer_pks at the beginning of each instruction execution_step.

RUN_PARAMS

Print out running arguments and params of a script.

SUBGRAPH_COMPLETE

Print out subgraph_complete value at the beginning of each instruction execution_step.

SUBGRAPH_ELEMENTS

Print out count of element in the current subgraph at the beginning of each instruction execution_step.

TARGET_MAP

This map should be used by rust-sdk logger that allows print only necessary targets by id.